#2 Implement DataCite client
