--- src/main/resources/kernel-4.1/include/datacite-relatedIdentifierType-v4.xsd
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Version 1.0 - Created 2011-01-13 - FZ, TIB, Germany
-  2013-05 v3.0: Addition of ID to simpleType element; addition of value "PMID"
-  2014-08-20 v3.1: Addition of values "arxiv" and "bibcode"
-  2015-02-12 v4.0 Addition of value "IGSN" -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://datacite.org/schema/kernel-4" targetNamespace="http://datacite.org/schema/kernel-4" elementFormDefault="qualified">
-  <xs:simpleType name="relatedIdentifierType" id="relatedIdentifierType">
-    <xs:annotation>
-      <xs:documentation>The type of the RelatedIdentifier.</xs:documentation>
-    </xs:annotation>
-    <xs:restriction base="xs:string">
-      <xs:enumeration value="ARK"/>
-      <xs:enumeration value="arXiv"/>
-      <xs:enumeration value="bibcode"/>
-      <xs:enumeration value="DOI"/>
-      <xs:enumeration value="EAN13"/>
-      <xs:enumeration value="EISSN"/>
-      <xs:enumeration value="Handle"/>
-      <xs:enumeration value="IGSN"/>
-      <xs:enumeration value="ISBN"/>
-      <xs:enumeration value="ISSN"/>
-      <xs:enumeration value="ISTC"/>
-      <xs:enumeration value="LISSN"/>
-      <xs:enumeration value="LSID"/>
-      <xs:enumeration value="PMID"/>
-      <xs:enumeration value="PURL"/>
-      <xs:enumeration value="UPC"/>
-      <xs:enumeration value="URL"/>
-      <xs:enumeration value="URN"/>
-    </xs:restriction>
-  </xs:simpleType>
-</xs:schema>
