--- src/main/resources/kernel-4.1/include/datacite-resourceType-v4.1.xsd
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Version 1.0 - Created 2011-01-13 - FZ, TIB, Germany
-     2013-05 v3.0: Addition of ID to simpleType element; added values "Audiovisual", "Workflow" & "Other"; deleted value "Film"
-     2017-10-23 v4.1: Addition of value "DataPaper" -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://datacite.org/schema/kernel-4" targetNamespace="http://datacite.org/schema/kernel-4" elementFormDefault="qualified">
-  <xs:simpleType name="resourceType" id="resourceType">
-    <xs:annotation>
-      <xs:documentation>The general type of a resource.</xs:documentation>
-    </xs:annotation>
-    <xs:restriction base="xs:string">
-      <xs:enumeration value="Audiovisual"/>
-      <xs:enumeration value="Collection"/>
-      <xs:enumeration value="DataPaper"/>
-      <xs:enumeration value="Dataset"/>
-      <xs:enumeration value="Event"/>
-      <xs:enumeration value="Image"/>
-      <xs:enumeration value="InteractiveResource"/>
-      <xs:enumeration value="Model"/>
-      <xs:enumeration value="PhysicalObject"/>
-      <xs:enumeration value="Service"/>
-      <xs:enumeration value="Software"/>
-      <xs:enumeration value="Sound"/>
-      <xs:enumeration value="Text"/>
-      <xs:enumeration value="Workflow"/>
-      <xs:enumeration value="Other"/>
-    </xs:restriction>
-  </xs:simpleType>
-</xs:schema>
