--- src/main/resources/kernel-4.2/include/datacite-funderIdentifierType-v4.xsd
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Version 1.0 - Created 2016-05-14 -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://datacite.org/schema/kernel-4" targetNamespace="http://datacite.org/schema/kernel-4" elementFormDefault="qualified">
-  <xs:simpleType name="funderIdentifierType" id="funderIdentifierType">
-    <xs:annotation>
-      <xs:documentation>The type of the funderIdentifier.</xs:documentation>
-    </xs:annotation>
-    <xs:restriction base="xs:string">
-      <xs:enumeration value="ISNI"/>
-      <xs:enumeration value="GRID"/>
-      <xs:enumeration value="Crossref Funder ID"/>
-      <xs:enumeration value="Other"/>
-    </xs:restriction>
-  </xs:simpleType>
-</xs:schema>
