--- src/main/resources/kernel-4.2/include/datacite-nameType-v4.xsd
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Version 4.1 - Created 2017-10-23 -->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://datacite.org/schema/kernel-4" targetNamespace="http://datacite.org/schema/kernel-4" elementFormDefault="qualified">
-  <xs:simpleType name="nameType" id="nameType">
-    <xs:restriction base="xs:string">
-      <xs:enumeration value="Organizational"/>
-      <xs:enumeration value="Personal"/>
-    </xs:restriction>
-  </xs:simpleType>
-</xs:schema>
